Modify

+++ EcoLokal/Models/Seller.swift
@@ -13,5 +13,5 @@ struct Seller: Identifiable, Codable {
     var category: String
     var city: String
     var description: String
-    var rating: Double = 0
+    var rating: Int = 0
 }
